Terminal: Fixed to compile with MSC 19 (Preparing to add initial Windows 10 native pty support)

+++ CtrlLib/Terminal/Sixel.h
@@ -33,7 +33,7 @@ private:
 private:
     Stream&         sstream;
     Vector<Color>   palette;
-    int             datapos;
+    int64           datapos;
     int             repeatcount;
     int             aspectratio;
     int             pass;
